635: (jal_purge) Create a list of records to be purged and then remove them instead of re-initiating the query after removing each record. 639: (jal-local-store) Do not delete jal.socket created by a previous instance. 636: (jal_purge) Return non-zero on error. 636: (jal_purge) Check if db_root exists and is a directory or return error. 629: (jaln_context) Replaced the defined *INIT* and *MAX* threads passed to apr thread call by apr_size_t variables. Strange, that seems to help! 629: (jald) Added a sleep(1) between jaln_context_destroy() and jaln_context_create(). 632, 633: (jald) Disconnect all sessions for a subscriber when one j/a/l session fails so the peer can reconnect afresh. 634: (jal_purge) Mark the records as "Deleted' or 'Kept' in the report of 'jal_purge' as appropriate.

+++ src/local_store/src/jalls.c
@@ -95,6 +95,7 @@ int main(int argc, char **argv) {
 	struct jalls_context *jalls_ctx = NULL;
 	enum jal_status jal_err = JAL_E_INVAL;
 	int sock = -1;
+	int old_socket_exist = 0;
 
 	if (0 != jalls_init()) {
 		goto err_out;
@@ -194,6 +195,8 @@ int main(int argc, char **argv) {
 	err = stat(jalls_ctx->socket, &sock_stat);
 	if (err != -1) {
 		fprintf(stderr, "failed to create socket: already exists\n");
+		fprintf(stderr, "Exiting ...\n");
+		old_socket_exist = 1;
 		goto err_out;
 	}
 	if (errno != ENOENT) {
@@ -279,7 +282,7 @@ int main(int argc, char **argv) {
 	}
 
 err_out:
-	if (jalls_ctx) {
+	if (jalls_ctx && 0 == old_socket_exist) {
 		delete_socket(jalls_ctx->socket, jalls_ctx->debug);
 	}
